Resolve "Loggerをjson to yamlする"

See merge request ricos/monolish!15

+++ src/blas/sparse/crs_spmv.cpp
@@ -14,7 +14,7 @@ namespace monolish{
 		logger.func_in(monolish_func);
 
 		//err
-		if( x.size() != y.size() || A.get_row() != (int)x.size()){
+		if( x.size() != y.size() || A.get_row() != (size_t)x.size()){
 			throw std::runtime_error("error vector size is not same");
 
 		}
@@ -34,13 +34,13 @@ namespace monolish{
 			#pragma acc kernels
 			{
 				#pragma acc loop independent 
-				for(int i = 0 ; i < n; i++){
+				for(size_t i = 0 ; i < n; i++){
 					yd[i] = 0;
 				}
 
 				#pragma acc loop independent
-				for(int i = 0 ; i < n; i++){
-					for(int j = rowd[i] ; j < rowd[i+1]; j++){
+				for(size_t i = 0 ; i < n; i++){
+					for(size_t j = rowd[i] ; j < rowd[i+1]; j++){
 						yd[i] += vald[j] * xd[cold[j]];
 					}
 				}
@@ -50,12 +50,12 @@ namespace monolish{
 #else // cpu
 
 	#pragma omp parallel for 
-		for(int i = 0 ; i < A.get_row(); i++)
+		for(size_t i = 0 ; i < A.get_row(); i++)
 			y.val[i] = 0;
 
 	#pragma omp parallel for
-		for(int i = 0 ; i < A.get_row(); i++)
-			for(int j = A.row_ptr[i] ; j < A.row_ptr[i+1]; j++)
+		for(size_t i = 0 ; i < A.get_row(); i++)
+			for(size_t j = (int)A.row_ptr[i] ; j < (int)A.row_ptr[i+1]; j++)
 				y.val[i] += A.val[j] * x.val[A.col_ind[j]];
 
 #endif
